requirements.txt
pytest==7.4.0
Flask==2.3.2
psycopg2-binary